+++ src/app/components/map/dialog.tsx
@@ -1,0 +1,7 @@
+export const Dialog = () => {
+	return (
+		<svg xmlns="http://www.w3.org/2000/svg" width="317" height="195" fill="none" viewBox="0 0 317 195">
+			<path fill="#D31346" stroke="#D31346" strokeWidth="3" d="m89.95 11.365-18.97 7.662-12.933 5.96-13.796 7.662-12.072 8.513-10.347 8.514-9.485 11.919-5.173 8.513-4.312 10.216L2 90.541l.862 13.621 4.312 12.77 5.173 10.217 9.485 11.067 10.347 9.365 12.072 8.514 13.796 6.81 12.934 5.109 12.934 5.108 13.796 4.256 12.934 2.554 17.245 2.554 18.969 1.703h32.766l16.383-1.703 16.383-2.554 14.659-4.256 12.933-5.108 10.348-2.554 15.52 5.108 12.072 6.81L296.893 191l-6.036-20.432-4.312-14.473v-10.217l4.312-4.256 8.622-10.217 8.623-10.216 4.311-10.216L315 100.757v-9.365l-2.587-15.324-4.311-10.217-8.623-12.77-11.209-11.068-10.347-7.662-13.796-8.513-18.108-8.514-15.52-5.96-18.108-4.256-16.383-2.554L171.865 2h-25.006l-28.454 2.554-28.455 6.81Z"/>
+		</svg>
+	);
+}
