+++ src/app/constants.tsx
@@ -5,3 +5,11 @@ export const LINKS = {
 	EMAIL: "twb.teachingwithoutborders@gmail.com", // TODO: Maybe we should offer to setup an email for them with their domain
 	UXPAND_LINKEDIN: "https://www.linkedin.com/company/uxpand-collective/"
 } as const;
+
+
+export const CountryInfo = {
+	BR: { text: "TODO: change it", center: [342, 213] },
+	SN: { text: "TODO: change it", center: [501, 112] },
+	MZ: { text: "TODO: change it", center: [787, 270] },
+	MG: { text: "TODO: change it", center: [837, 308] },
+};
